Updated names and added autofac DI
