+++ AngularApp.Core/Interfaces/Services/ITeacherService.cs
@@ -1,0 +1,12 @@
+﻿using Core.Models;
+
+namespace Core.Interfaces.Services
+{
+	public interface ITeacherService
+	{
+		int Save(Teacher teacherToSave);
+		void Update(Teacher teacherToUpdate);
+		Teacher Get(int teacherId);
+		void Delete(int teacherId);
+	}
+}
