+++ AngularApp.Core/Models/Class.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Core.Models
+{
+	public class Class
+	{	
+		public int InstanceId { get; set; }
+		public Teacher Teacher { get; set; }
+		public IEnumerable<Student> Students { get; set; }
+	}
+}
