+++ AngularApp.Core/Models/SlimClass.cs
@@ -1,0 +1,9 @@
+﻿namespace Core.Models
+{
+	public class SlimClass
+	{
+		public int InstanceId { get; set; }
+		public int TeacherId { get; set; }
+		public int StudentId { get; set; }
+	}
+}
