+++ AngularApp.Infrastructure/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Dapper" version="1.42" targetFramework="net45" />
+</packages>
