--- Core/Interfaces/Repositories/IClassRepository.cs
@@ -1,15 +1,0 @@
-﻿using System.Collections.Generic;
-using Core.Models;
-
-namespace Core.Interfaces.Repositories
-{
-	public interface IClassRepository
-	{
-		int Create(SlimClass insert);
-		void SaveTeacher(int instanceId, int teacherId);
-		void SaveStudent(SlimClass insert);
-		IEnumerable<SlimClass> Get(int instanceId);
-		void Delete(int instanceId);
-		void DeleteStudent(int instanceId, int studentId);
-	}
-}
