--- Core/Interfaces/Repositories/ITeacherRepository.cs
@@ -1,12 +1,0 @@
-﻿using Core.Models;
-
-namespace Core.Interfaces.Repositories
-{
-	public interface ITeacherRepository
-	{
-		int Save(Teacher teacherToSave);
-		void Update(Teacher teacherToUpdate);
-		Teacher Get(int teacherId);
-		void Delete(int teacherId);
-	}
-}
