--- Core/Models/SlimClass.cs
@@ -1,9 +1,0 @@
-﻿namespace Core.Models
-{
-	public class SlimClass
-	{
-		public int InstanceId { get; set; }
-		public int TeacherId { get; set; }
-		public int StudentId { get; set; }
-	}
-}
