--- Infrastructure/Services/StudentService.cs
@@ -1,36 +1,0 @@
-﻿using Core.Interfaces.Repositories;
-using Core.Interfaces.Services;
-using Core.Models;
-
-namespace Infrastructure.Services
-{
-	public class StudentService : IStudentService
-	{
-		private readonly IStudentRepository _repository;
-
-		public StudentService(IStudentRepository repository)
-		{
-			_repository = repository;
-		}
-
-		public int Save(Student studentToSave)
-		{
-			return _repository.Save(studentToSave);
-		}
-
-		public void Update(Student studentToUpdate)
-		{
-			_repository.Update(studentToUpdate);			
-		}
-
-		public Student Get(int studentId)
-		{
-			return _repository.Get(studentId);
-		}
-
-		public void Delete(int studentId)
-		{
-			_repository.Delete(studentId);
-		}
-	}
-}
